requirements.txt
simpy
pandas
boto3
requests
python-dotenv
